recommender/requirements.txt
google-generativeai>=0.3.0
firebase-admin>=6.4.0
python-dotenv>=1.0.1
